Replace the new backend in start command
