+++ src/engine/ruleset/wazuh-core/manifest.yml
@@ -29,9 +29,6 @@ decoders:
   - decoder/upgrade/0
   - decoder/core-windows/0
 
-filters:
-  - filter/allow-all/0
-
 outputs:
   - output/file-output-wazuh-core/0
   - output/file-output-integrations/0
