+++ src/engine/source/router/src/runtimePolicy.cpp
@@ -1,9 +1,9 @@
 #include <router/runtimePolicy.hpp>
 
-#include <base/parseEvent.hpp>
+#include <parseEvent.hpp>
 #include <logging/logging.hpp>
-#include <rxbk/rxFactory.hpp>
 #include <utils/getExceptionStack.hpp>
+#include <bk/taskf/controller.hpp>
 
 namespace router
 {
@@ -27,7 +27,14 @@ std::optional<base::Error> RuntimePolicy::build(std::shared_ptr<builder::Builder
             return base::Error {fmt::format("Policy '{}' has no assets", m_asset)};
         }
 
-        m_controller = std::make_shared<rxbk::Controller>(rxbk::buildRxPipeline(policy));
+        // TODO Check de assets names policy api (Return a string instead of a base::Names?)
+        std::unordered_set<std::string> assetNames;
+        std::transform(policy.assets().begin(),
+                       policy.assets().end(),
+                       std::inserter(assetNames, assetNames.begin()),
+                       [](const auto& name) { return name.toStr(); });
+
+        m_controller = std::make_shared<bk::taskf::Controller>(policy.expression(), assetNames);
         m_hash = policy.hash();
     }
     catch (std::exception& e)
@@ -38,32 +45,38 @@ std::optional<base::Error> RuntimePolicy::build(std::shared_ptr<builder::Builder
     return std::nullopt;
 }
 
-std::optional<base::Error> RuntimePolicy::processEvent(base::Event event)
+std::optional<base::Error> RuntimePolicy::processEvent(base::Event&& event)
 {
     if (!m_controller)
     {
         return base::Error {fmt::format("Policy '{}' is not built", m_asset)};
     }
-    auto result = base::result::makeSuccess(event);
-    m_controller->ingestEvent(std::make_shared<base::result::Result<base::Event>>(std::move(result)));
+
+    if (m_trace.publishOutput) {
+        event = m_controller->ingestGet(std::move(event));
+        if (event) {
+            m_trace.publishOutput(std::move(event));
+        }
+    }
+    else {
+        m_controller->ingest(std::move(event));
+    }
 
     return std::nullopt;
 }
 
-std::optional<base::Error> RuntimePolicy::subscribeToOutput(rxbk::SubscribeToOutputCallback callback)
+std::optional<base::Error> RuntimePolicy::subscribeToOutput(const OutputSubscriber& callback)
 {
     if (!callback)
     {
         return base::Error {SUBSCRIBE_CONFIGURATION_ERROR};
     }
-
-    auto subscriber = m_controller->configureSuscribeToOutput(callback);
-    m_csOutput = m_controller->getOutput().subscribe(std::move(subscriber));
+    m_trace.publishOutput = callback;
 
     return std::nullopt;
 }
 
-std::optional<base::Error> RuntimePolicy::listenAllTrace(rxbk::SubscribeToTraceCallback callback,
+std::optional<base::Error> RuntimePolicy::listenAllTrace(const bk::Subscriber& callback,
                                                          const std::vector<std::string>& assets)
 {
     if (!callback)
@@ -71,35 +84,47 @@ std::optional<base::Error> RuntimePolicy::listenAllTrace(rxbk::SubscribeToTraceC
         return base::Error {SUBSCRIBE_CONFIGURATION_ERROR};
     }
 
-    auto subscriber = m_controller->configureSuscribeToTrace(callback);
+    std::vector<std::pair< std::string, bk::Subscription>> subscriptions {};
+    subscriptions.reserve(assets.size());
 
-    try
+    for (const auto& asset : assets)
     {
-        m_csTraces = m_controller->listenOnAllTrace(std::move(subscriber), assets);
+        auto res = m_controller->subscribe(asset, callback);
+        if (base::isError(res))
+        {
+            return base::Error {base::getError(res).message};
+        }
+        subscriptions.emplace_back(asset, base::getResponse(res));
     }
-    catch (const std::exception& e)
-    {
-        return base::Error {e.what()};
-    }
+    m_trace.subscriptions = std::move(subscriptions);
 
     return std::nullopt;
 }
 
-const std::vector<std::string> RuntimePolicy::getAssets() const
+std::vector<std::string> RuntimePolicy::getAssets() const
 {
-    return m_controller->getAssets();
+    // TODO CHANGE THIS
+    // Temp fix to tipes
+
+    const auto& setAsset = m_controller->getTraceables();
+
+    std::vector<std::string> assets {};
+    assets.reserve(setAsset.size());
+    std::transform(
+        setAsset.begin(), setAsset.end(), std::back_inserter(assets), [](const auto& asset) { return asset; });
+
+    return assets;
 }
 
 void RuntimePolicy::unSubscribeTraces()
 {
-    if (m_csTraces.is_subscribed())
+    for (const auto& [asset, subscription] : m_trace.subscriptions)
     {
-        m_csTraces.unsubscribe();
+        m_controller->unsubscribe(asset, subscription);
     }
-    if (m_csOutput.is_subscribed())
-    {
-        m_csOutput.unsubscribe();
-    }
+
+    m_trace.subscriptions.clear();
+    m_trace.publishOutput = nullptr;
 }
 
 } // namespace router
