+++ src/engine/tools/vagrant/Vagrantfile
@@ -149,6 +149,9 @@ Vagrant.configure("2") do |config|
         engine-integration add -n wazuh integrations/apache-http/
         engine-integration add -n wazuh integrations/suricata/
 
+        # Add filter for route
+        $ENGINE_DIR/wazuh-engine catalog -n system create filter << filters/allow-all.yml
+
         cd $HOME
         echo "Creating default policy 'policy/wazuh/0'..."
         $ENGINE_DIR/wazuh-engine policy add -p policy/wazuh/0 -f
